requirements.txt
# Core ETL tools
feedparser>=6.0.10
python-dateutil>=2.8.2

# NLP
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

streamlit>=1.35.0
pandas>=2.2.2


scikit-learn>=1.3.0

APScheduler>=3.10.4

